\name{Cadastral2}
\alias{Cadastral2}
\docType{data}
\title{
Cadastral Units (CU) Level 2
}
\description{
Cadastral Units (CU) Level 2
}
\usage{data("Cadastral2")}
\details{
Dataset contains following columns:

id - denoting plot order

name - denoting name of CU2 unit

code - denote LAU2 code (first six numbers denote LAU1 code)

coord_x, coord_y - denote coordinates of centroids
}
\source{
Source of maps: \url{http://www.codgik.gov.pl/index.php/darmowe-dane/prg.html}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Cadastral1)
## maybe plot(Cadastral1) ...
}
\keyword{datasets}
