% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_ggmap.R
\name{fast_ggmap}
\alias{fast_ggmap}
\title{Function fast_ggmap}
\usage{
fast_ggmap(data, id, var, maptype = "powiat", higher_level = FALSE)
}
\arguments{
\item{higher_level}{}
}
\value{
ggplot2
}
\description{
This function plots maps, very fast
}
\examples{
# toDo
}
\keyword{ggplot2}
\keyword{sp,}

