\name{LAU1}
\alias{LAU1}
\docType{data}
\title{
The Local Administrative Units (LAU) Level 1
}
\description{
The Local Administrative Units (LAU) Level 1 – powiats and cities with powiat status
}
\usage{data("LAU1")}
\details{
Dataset contains following columns:

id - denoting plot order

name - denoting name of LAU1 unit

code - denote LAU1 code

coord_x, coord_y - denote coordinates of centroids

}
\source{
Poland: \url{http://stat.gov.pl/en/regional-statistics/nomenclature-nts-161/}

Eurostat: \url{http://ec.europa.eu/eurostat/web/nuts/local-administrative-units}

Source of maps: \url{http://www.codgik.gov.pl/index.php/darmowe-dane/prg.html}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(LAU1)
## maybe str(LAU1) ; plot(LAU1) ...
}
\keyword{datasets}
