\name{LAU2}
\alias{LAU2}
\docType{data}
\title{
The Local Administrative Units (LAU) Level 2
}
\description{
The Local Administrative Units (LAU) Level 2 – gminas, of which urban gminas that are cities with powiat status
}
\usage{data("LAU2")}
\details{
Dataset contains following columns:

id - denoting plot order

name - denoting name of LAU2 unit

code - denote LAU2 code

coord_x, coord_y - denote coordinates of centroids
}
\source{
Poland: \url{http://stat.gov.pl/en/regional-statistics/nomenclature-nts-161/}

Eurostat: \url{http://ec.europa.eu/eurostat/web/nuts/local-administrative-units}

Source of maps: \url{http://www.codgik.gov.pl/index.php/darmowe-dane/prg.html}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(LAU2)
## maybe str(LAU2) ; plot(LAU2) ...
}
\keyword{datasets}
