% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spdf_to_df.R
\name{spdf_to_df}
\alias{spdf_to_df}
\title{SpatialPolygonsDataFrame to data.frame function}
\usage{
spdf_to_df(spdf)
}
\arguments{
\item{spdf}{- SpatialPolygonsDataFrame}
}
\value{
data.frame
}
\description{
This function convert SpatialPolygonsDataFrame to data.frame witout droping any variable
}
\examples{
# data("wojewdztwa")
# spdf_to_df(wojewodztwa)
}
\keyword{ggplot2}
\keyword{sp,}

