% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_ggmap.R
\name{fast_ggmap}
\alias{fast_ggmap}
\title{Simple function to plot data on map of Poland}
\usage{
fast_ggmap(data, id_var, fill_var, maptype = "powiat", higher_level = T,
  viridis = T, ...)
}
\arguments{
\item{data}{- data.frame with id and variables for plot}

\item{id_var}{- id for spatial units}

\item{fill_var}{- variable to be plotted}

\item{maptype}{- administrative level}

\item{higher_level}{- whether to include higher level}

\item{viridis}{- whether to use}

\item{...}{-dots}
}
\value{
ggplot2
}
\description{
This function
}
\examples{
# data(lau2)
#
}
\keyword{map}

